Adding tree generation from an array for C language
Adding code to make a binary tree from an array input. The code is
in C language. The reason is to add C language solutions to the
Leetcode problems. Also added unit testing for C using the
Ceedling/Unity framework but did not add the configuration files
for the framework at this time.